backend/pyproject.toml
[project]
name = "gym-pr-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi[standard]>=0.121.3",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
    "sqlmodel>=0.0.32",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.1.0",
    "rq>=2.6.1",
    "google-generativeai>=0.8.6",
    "httpx>=0.27.0",
    "openai>=1.14.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "ruff>=0.14.8",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
